requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Werkzeug==2.2.3
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1
psutil==5.9.8
Pillow==10.2.0
cachetools==5.3.2